fix(review-round-9): eliminate as/unwrap/expect risks in core crates
- Replace silent  casts with fallible  conversions or  where available.
- Paginate with  from ; remove  in repositories.
- Convert upload-session part counts, dyun heartbeat timeout, lease duration,
  rate-limit remaining tokens and reset durations, and fencing timestamps safely.
- Derive  for  and compare directly instead of .
- Allow  conversions only in generated protobuf code, trait-object coercion,
  and guarded f64 coordinate helpers.
- Add  poison recovery and  redirect disabling from round 8.

+++ crates/domain/src/hardware.rs
@@ -198,7 +198,7 @@ impl HardwareCompatibility {
         }
 
         let tier = resource_class.support_tier;
-        if tier as i32 > requirement.max_support_tier as i32 {
+        if tier > requirement.max_support_tier {
             return Compatibility::Blocked;
         }
